Move 'ConstExtract' request from TypeCheckRequest into its own component's collection of requests.
