+++ include/swift/AST/ASTTypeIDs.h
@@ -82,6 +82,7 @@ enum class ImplicitMemberAction : uint8_t;
 struct FingerprintAndMembers;
 class Identifier;
 class BodyAndFingerprint;
+struct ConstValueTypeInfo;
 
 // Define the AST type zone (zone 1)
 #define SWIFT_TYPEID_ZONE AST
