+++ include/swift/Basic/TypeIDZones.def
@@ -35,5 +35,7 @@ SWIFT_TYPEID_ZONE(IDE, 137)
 
 SWIFT_TYPEID_ZONE(ClangImporter, 139)
 
+SWIFT_TYPEID_ZONE(ConstExtract, 140)
+
 // N.B. This is not a formal zone and exists solely to support the unit tests.
 SWIFT_TYPEID_ZONE(ArithmeticEvaluator, 255)
