+++ include/swift/ConstExtract/ConstExtractTypeIDZone.def
@@ -1,0 +1,20 @@
+//===--- ConstExtractTypeIDZone.def ----------------------------*- C++ -*-===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2021 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+//
+//  This definition file describes the types in the const extract
+//  TypeID zone, for use with the TypeID template.
+//
+//===----------------------------------------------------------------------===//
+
+SWIFT_REQUEST(ConstExtract, ConstantValueInfoRequest,
+              ConstValueTypeInfo(NominalTypeDecl *), Cached,
+              NoLocationInfo)
