+++ include/swift/Subsystems.h
@@ -386,6 +386,9 @@ namespace swift {
   /// after forming the ASTContext.
   void registerClangImporterRequestFunctions(Evaluator &evaluator);
 
+  /// Register constant value extraction request functons with the evaluator.
+  void registerConstExtractRequestFunctions(Evaluator &evaluator);
+
   /// Register SILOptimizer passes necessary for IRGen.
   void registerIRGenSILTransforms(ASTContext &ctx);
